requirements.txt
boto3==1.26.94
botocore==1.29.94
click==8.1.3
cmake==3.25.0
colorama==0.4.6
dlib==19.9.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
numpy==1.24.1
opencv-contrib-python==4.7.0.72
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.2
zipp==3.15.0
